refactor(memory): rehome host namespace store
