+++ src/openhuman/memory/read_rpc_tests.rs
@@ -4,7 +4,7 @@ use crate::openhuman::embeddings::NoopEmbedding;
 use crate::openhuman::memory::ingest_pipeline::ingest_chat;
 use crate::openhuman::memory_queue::drain_until_idle;
 use crate::openhuman::memory_store::content::raw::{write_raw_items, RawItem, RawKind};
-use crate::openhuman::memory_store::unified::UnifiedMemory;
+use crate::openhuman::memory_store::namespace_store::UnifiedMemory;
 use crate::openhuman::memory_sync::canonicalize::chat::{ChatBatch, ChatMessage};
 use chrono::{TimeZone, Utc};
 use rusqlite::params;
