+++ src/openhuman/memory_store/mod.rs
@@ -11,7 +11,8 @@
 //! ## Submodules
 //!
 //! - `types`: Common data structures and types used across the memory store.
-//! - `unified`: The primary SQLite-based memory implementation.
+//! - `namespace_store`: Host-retained SQLite namespace documents, graph,
+//!   episodic/event/segment/profile tables, and their query policy.
 //! - `client`: High-level client interface for interacting with the memory system.
 //! - `factories`: Factory functions for creating and initializing memory instances.
 //! - `memory_trait`: Defines the `Memory` trait that all implementations must satisfy.
@@ -21,13 +22,13 @@ pub mod content;
 pub mod entities;
 pub mod kinds;
 pub mod kv;
+pub mod namespace_store;
 pub mod retrieval;
 pub mod safety;
 pub mod tools;
 pub mod traits;
 pub mod trees;
 pub mod types;
-pub mod unified;
 pub mod vectors;
 
 mod client;
@@ -42,16 +43,16 @@ pub use factories::{
     active_embedding_signature, create_memory, create_memory_for_migration,
     create_memory_with_local_ai, effective_embedding_settings, effective_memory_backend_name,
 };
+pub use namespace_store::events;
+pub use namespace_store::fts5;
+pub use namespace_store::profile;
+pub use namespace_store::segments;
+pub use namespace_store::UnifiedMemory;
 pub use types::{
     GraphRelationRecord, MemoryItemKind, MemoryKvRecord, NamespaceDocumentInput,
     NamespaceMemoryHit, NamespaceQueryResult, NamespaceRetrievalContext, RetrievalScoreBreakdown,
     StoredMemoryDocument,
 };
-pub use unified::events;
-pub use unified::fts5;
-pub use unified::profile;
-pub use unified::segments;
-pub use unified::UnifiedMemory;
 
 #[cfg(test)]
 mod tests {
