+++ src/openhuman/memory_store/namespace_store/events.rs
