+++ src/openhuman/memory_store/namespace_store/profile.rs
