+++ src/openhuman/memory_store/namespace_store/segments.rs
